+++ version_history.md
@@ -3,6 +3,7 @@
 ## 1.1.0 - TBD
 
 * Made `constexpr_static_assert` constexpr.
+* Added default user and channel to conanfile.
 
 ## 1.0.0 - April 2023